+++ choco-solver/src/main/java/org/chocosolver/solver/search/solution/LastSolutionRecorder.java
@@ -62,7 +62,7 @@ public class LastSolutionRecorder implements ISolutionRecorder {
 						try{
 							solver.getSearchLoop().restoreRootNode();
 							solver.getEnvironment().worldPush();
-							solution.restore();
+							solution.restore(solver);
 						}catch (ContradictionException e){
 							throw new UnsupportedOperationException("restoring the last solution ended in a failure");
 						}